+++ brewtils/auto_decorator.py
@@ -1,7 +1,7 @@
 from inspect import Parameter as InspectParameter  # noqa
 from inspect import signature
 
-from brewtils.models import Command, Parameter
+from brewtils.models import Command
 
 
 class AutoDecorator: